+++ plugin/web/extensions/sidebar/core/components/content/common/DatasetCard/Field/Fields/3dtiles/BuildingTransparency/useBuildingTransparency.ts
@@ -1,5 +1,7 @@
-import { getTransparencyExpression } from "@web/extensions/sidebar/utils/color";
-import { getOverriddenLayerByDataID } from "@web/extensions/sidebar/utils/getOverriddenLayerByDataID";
+import {
+  getTransparencyExpression,
+  getOverriddenLayerByDataID,
+} from "@web/extensions/sidebar/utils";
 import debounce from "lodash/debounce";
 import { MutableRefObject, RefObject, useCallback, useEffect, useMemo, useRef } from "react";
